+++ src/app/modules/search-parameter/search-parameter.component.ts
@@ -8,7 +8,6 @@ import {
   createControlValueAccessorProviders
 } from '../base-control-value-accessor';
 import { FhirBackendService } from '../../shared/fhir-backend/fhir-backend.service';
-import { SearchCondition } from '../../types/search.condition';
 import {
   encodeFhirSearchParameter,
   escapeFhirSearchParameter
@@ -26,15 +25,18 @@ import {
 export class SearchParameterComponent
   extends BaseControlValueAccessor<SearchParameter>
   implements OnInit {
-  @Input() inputResourceType = '';
-  fixedResourceType = false;
-  readonly OBSERVATIONBYTEST = 'Observation by Test';
+  @Input() resourceType = '';
+  readonly OBSERVATIONBYTEST = 'code text';
   readonly CODETYPES = ['code', 'CodeableConcept', 'Coding'];
+  // Observation search parameter names to be hidden
+  readonly OBSERVATIONHIDDENPARAMETERNAMES = [
+    'combo-code',
+    'combo-value-concept',
+    'combo-value-quantity',
+    'value-string'
+  ];
   definitions: any;
 
-  resourceType: FormControl = new FormControl('');
-  resourceTypes: string[] = [];
-  filteredResourceTypes: Observable<string[]>;
   selectedResourceType: any;
 
   parameterName: FormControl = new FormControl('');
@@ -49,7 +51,6 @@ export class SearchParameterComponent
 
   get value(): SearchParameter {
     return {
-      resourceType: this.resourceType.value,
       name: this.parameterName.value,
       value: this.parameterValue.value,
       selectedObservationCodes: this.selectedObservationCodes.value
@@ -72,44 +73,18 @@ export class SearchParameterComponent
   }
 
   ngOnInit(): void {
-    this.filteredResourceTypes = this.resourceType.valueChanges.pipe(
-      startWith(''),
-      map((value) => this._filter(value, this.resourceTypes))
-    );
-
-    this.resourceType.valueChanges.subscribe((value) => {
-      if (value === this.OBSERVATIONBYTEST) {
-        this.selectedParameter = null;
-        this.selectedResourceType = null;
-        return;
-      }
-      const match = this.resourceTypes.find((rt) => rt === value);
-      if (match) {
-        this.selectedResourceType = this.definitions.resources[value];
-        this.parameterNames = this.selectedResourceType.searchParameters.map(
-          (sp) => sp.name
-        );
-        this.selectedParameter = null;
-        this.parameterName.setValue('');
-        this.parameterValue.setValue('');
-      }
-      // tell Angular forms API to update parent form control
-      this.onChange(this.value);
-    });
-
     this.definitions = this.fhirBackend.getCurrentDefinitions();
-    if (!this.inputResourceType) {
-      this.resourceTypes = Object.keys(this.definitions.resources).concat(
-        this.OBSERVATIONBYTEST
+    this.selectedResourceType = this.definitions.resources[this.resourceType];
+    this.parameterNames = this.selectedResourceType.searchParameters.map(
+      (sp) => sp.name
+    );
+    if (this.resourceType === 'Observation') {
+      this.parameterNames = this.parameterNames.filter(
+        (n) => !this.OBSERVATIONHIDDENPARAMETERNAMES.includes(n)
       );
-    } else if (this.inputResourceType === 'Observation') {
-      this.resourceTypes = ['Observation', this.OBSERVATIONBYTEST];
-    } else {
-      // single resource type
-      this.resourceType.setValue(this.inputResourceType);
-      this.resourceTypes = [this.inputResourceType];
-      this.fixedResourceType = true;
+      this.parameterNames.unshift(this.OBSERVATIONBYTEST);
     }
+    this.selectedParameter = null;
 
     this.filteredParameterNames = this.parameterName.valueChanges.pipe(
       startWith(''),
@@ -117,17 +92,15 @@ export class SearchParameterComponent
     );
 
     this.parameterName.valueChanges.subscribe((value) => {
-      if (this.selectedResourceType) {
-        this.selectedParameter = this.selectedResourceType.searchParameters.find(
-          (p) => p.name === value
-        );
-        if (this.selectedParameter) {
-          this.parameterValue.setValue('');
-          if (this.selectedParameter.valueSet) {
-            this.parameterValues = this.definitions.valueSets[
-              this.selectedParameter.valueSet
-            ];
-          }
+      this.selectedParameter = this.selectedResourceType.searchParameters.find(
+        (p) => p.name === value
+      );
+      if (this.selectedParameter) {
+        this.parameterValue.setValue('');
+        if (this.selectedParameter.valueSet) {
+          this.parameterValues = this.definitions.valueSets[
+            this.selectedParameter.valueSet
+          ];
         }
       }
       this.onChange(this.value);
@@ -162,7 +135,6 @@ export class SearchParameterComponent
    * @param value New value to be written to the model.
    */
   writeValue(value: SearchParameter): void {
-    this.resourceType.setValue(value.resourceType || '');
     this.parameterName.setValue(value.name || '');
     this.parameterValue.setValue(value.value || '');
     this.selectedObservationCodes.setValue(
@@ -171,24 +143,15 @@ export class SearchParameterComponent
   }
 
   /**
-   * return resource type and url segment of search string for current search parameter.
-   */
-  getCondition(): SearchCondition {
-    const criteria = this.getCriteria();
-    return criteria
-      ? {
-          resourceType: this.resourceType.value,
-          criteria
-        }
-      : null;
-  }
-
-  /**
    * get string of url segment describing the search criteria that will be used to search in server.
    */
   getCriteria(): string {
-    if (this.resourceType.value === this.OBSERVATIONBYTEST) {
+    if (this.parameterName.value === this.OBSERVATIONBYTEST) {
       return this.getObservationByTestCriteria();
+    }
+    // Return empty if parameter name is not selected.
+    if (!this.selectedParameter) {
+      return '';
     }
     if (this.selectedParameter.type === 'date') {
       return (
@@ -201,7 +164,7 @@ export class SearchParameterComponent
       );
     }
     if (
-      this.resourceType.value === 'Patient' &&
+      this.resourceType === 'Patient' &&
       this.parameterName.value === 'active' &&
       this.parameterValue.value === 'true'
     ) {
@@ -217,6 +180,10 @@ export class SearchParameterComponent
   }
 
   getObservationByTestCriteria(): string {
+    // Ignore criteria if no code selected.
+    if (!this.selectedObservationCodes.value) {
+      return '';
+    }
     const comboCodes = this.selectedObservationCodes.value.codes.filter(
       (c) => c
     );
@@ -237,17 +204,11 @@ export class SearchParameterComponent
         )
       : '';
     const unit = this.parameterValue.value.testValueUnit;
-    const from = this.parameterValue.value.from
-      ? `&date=ge${encodeURIComponent(this.parameterValue.value.from)}`
-      : '';
-    const to = this.parameterValue.value.to
-      ? `&date=le${encodeURIComponent(this.parameterValue.value.to)}`
-      : '';
     const valueParam = testValue.trim()
       ? `&${valueParamName}${modifier}=${prefix}${encodeURIComponent(
           testValue + (unit ? '||' + escapeFhirSearchParameter(unit) : '')
         )}`
       : '';
-    return `${codeParam}${valueParam}${from}${to}`;
+    return `${codeParam}${valueParam}`;
   }
 }