+++ src/app/modules/search-parameters/search-parameters.component.ts
@@ -1,12 +1,12 @@
-import { Component, Input, ViewChildren, QueryList } from '@angular/core';
+import { Component, ViewChildren, QueryList } from '@angular/core';
 import { AbstractControl, FormArray, FormControl } from '@angular/forms';
 import {
   BaseControlValueAccessor,
   createControlValueAccessorProviders
 } from '../base-control-value-accessor';
 import { SearchParameter } from 'src/app/types/search.parameter';
-import { SearchParameterComponent } from '../search-parameter/search-parameter.component';
 import { SearchCondition } from '../../types/search.condition';
+import { SearchParameterGroupComponent } from '../search-parameter-group/search-parameter-group.component';
 
 /**
  * Component for managing resources search parameters
@@ -20,72 +20,55 @@ import { SearchCondition } from '../../types/search.condition';
 export class SearchParametersComponent extends BaseControlValueAccessor<
   SearchParameter[]
 > {
-  /**
-   * Limits the list of available search parameters to only parameters for this resource type
-   */
-  @Input() resourceType = '';
-  @ViewChildren(SearchParameterComponent)
-  searchParameterComponents: QueryList<SearchParameterComponent>;
-  parameterList = new FormArray([]);
-  readonly OBSERVATIONBYTEST = 'Observation by Test';
+  @ViewChildren(SearchParameterGroupComponent)
+  searchParameterGroupComponents: QueryList<SearchParameterGroupComponent>;
+  parameterGroupList = new FormArray([]);
 
   constructor() {
     super();
-    this.parameterList.valueChanges.subscribe((value) => {
+    this.parameterGroupList.valueChanges.subscribe((value) => {
       this.onChange(value);
     });
   }
 
   /**
-   * Add new search parameter to search parameter list
+   * Add new search parameter group to search parameter group list
    */
-  public addParameter(): void {
-    this.parameterList.push(
+  public addParameterGroup(): void {
+    this.parameterGroupList.push(
       new FormControl({
-        resourceType: this.resourceType
+        resourceType: '',
+        parameters: []
       })
     );
   }
 
   /**
-   * Remove search parameter from search parameter list
+   * Remove search parameter group from search parameter group list
    */
-  public removeParameter(item: AbstractControl): void {
-    this.parameterList.removeAt(this.parameterList.controls.indexOf(item));
+  public removeParameterGroup(item: AbstractControl): void {
+    this.parameterGroupList.removeAt(
+      this.parameterGroupList.controls.indexOf(item)
+    );
   }
 
   writeValue(value: SearchParameter[]): void {
     // TODO
   }
 
-  // Get search conditions from each row, group them on the resource type
+  // Get search conditions from each row
   getConditions(): SearchCondition[] {
-    const conditions = this.searchParameterComponents.map((c) =>
-      c.getCondition()
-    );
-    const groupedConditions = [];
-    // add default Patient condition if missing
-    groupedConditions.push({
-      resourceType: 'Patient',
-      criteria: ''
-    });
-    conditions.forEach((item) => {
-      const match = groupedConditions.find(
-        (x) => x.resourceType === item.resourceType
-      );
-      // do not combine conditions for 'Observation by Test'
-      if (match && item.resourceType !== this.OBSERVATIONBYTEST) {
-        match.criteria += item.criteria;
-      } else {
-        groupedConditions.push(item);
-      }
-    });
-    conditions.map((item) => {
-      // for 'Observation by Test', search url needs to use 'Observation'
-      if (item.resourceType === this.OBSERVATIONBYTEST) {
-        item.resourceType = 'Observation';
-      }
-    });
-    return groupedConditions;
+    const conditions = this.searchParameterGroupComponents
+      .map((c) => c.getConditions())
+      // Filter out empty resource type or criteria
+      .filter((c) => c.resourceType && c.criteria);
+    if (!conditions.some((c) => c.resourceType === 'Patient')) {
+      // add default Patient condition if missing
+      conditions.push({
+        resourceType: 'Patient',
+        criteria: ''
+      });
+    }
+    return conditions;
   }
 }