+++ src/app/types/search.parameter.ts
@@ -1,5 +1,4 @@
 export interface SearchParameter {
-  resourceType?: string;
   name?: string;
   // TODO: value type TBD
   value?: any;